+++ android/CHANGELOG.md
@@ -24,6 +24,7 @@ Line wrap the file at 100 chars.                                              Th
 ## [Unreleased]
 ### Added
 - Add link to notifications system settings from notifications screen
+- Added support for account deletion
 
 ### Changed
 - Hide password by default on login screen.